Enable rayon by default and refactor parallel iter
