+++ src/lib.rs
@@ -161,7 +161,6 @@ extern crate num_traits;
 #[macro_use]
 extern crate cfg_if;
 extern crate im;
-#[cfg(feature = "rayon")]
 extern crate rayon;
 
 cfg_if! {
