+++ tests/parallel_tests.rs
@@ -1,9 +1,7 @@
-#![cfg(feature = "rayon")]
-
 extern crate generational_arena_im;
 extern crate rayon;
 use generational_arena_im::StandardArena as Arena;
-use rayon::iter::{IntoParallelRefIterator, IntoParallelRefMutIterator, ParallelIterator};
+use rayon::iter::{IntoParallelIterator, ParallelIterator};
 
 #[test]
 fn par_iter_matches_sequential() {
@@ -12,7 +10,7 @@ fn par_iter_matches_sequential() {
         arena.insert(i);
     }
     let seq: Vec<_> = arena.iter().map(|(_, v)| *v).collect();
-    let par: Vec<_> = arena.par_iter().map(|(_, v)| *v).collect();
+    let par: Vec<_> = (&arena).into_par_iter().map(|(_, v)| *v).collect();
 
     assert_eq!(seq, par);
 }
@@ -23,7 +21,7 @@ fn par_iter_mut_updates() {
     for i in 0..100 {
         arena.insert(i);
     }
-    arena.par_iter_mut().for_each(|(_, v)| *v *= 2);
+    (&mut arena).into_par_iter().for_each(|(_, v)| *v *= 2);
     let values: Vec<_> = arena.iter().map(|(_, v)| *v).collect();
     assert_eq!(values, (0..100).map(|x| x * 2).collect::<Vec<_>>());
 }
